+++ dflib/src/test/java/org/dflib/exp/flow/IfNullExpTest.java
@@ -2,14 +2,13 @@ package org.dflib.exp.flow;
 
 import org.dflib.DataFrame;
 import org.dflib.Exp;
-import org.dflib.exp.ExpBaseTest;
+import org.dflib.exp.BaseExpTest;
 import org.dflib.unit.SeriesAsserts;
 import org.junit.jupiter.api.Test;
 
 import static org.dflib.Exp.*;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-public class IfNullExpTest extends ExpBaseTest {
+public class IfNullExpTest extends BaseExpTest {
 
     @Test
     public void string() {
@@ -51,15 +50,28 @@ public class IfNullExpTest extends ExpBaseTest {
     }
 
     @Test
-    public void equalsHashCode() {
-        Exp<Integer> e1 = ifNull($int("a"), $int(1));
-        Exp<Integer> e2 = ifNull($int("a"), $int(1));
-        Exp<Integer> e3 = ifNull($int("a"), $int(1));
-        Exp<Integer> different1 = ifNull($int("b"), $int(1));
-        Exp<Integer> different2 = ifNull($int("a"), $int(2));
+    public void testEquals() {
+        assertExpEquals(
+                ifNull($int("a"), $int(1)),
+                ifNull($int("a"), $int(1)),
+                ifNull($int("b"), $int(1)));
 
-        assertEqualsContract(e1, e2, e3);
-        assertNotEquals(e1, different1);
-        assertNotEquals(e1, different2);
+        assertExpEquals(
+                ifNull($int("a"), $int(1)),
+                ifNull($int("a"), $int(1)),
+                ifNull($int("a"), $int(2)));
+    }
+
+    @Test
+    public void testHashCode() {
+        assertExpHashCode(
+                ifNull($int("a"), $int(1)),
+                ifNull($int("a"), $int(1)),
+                ifNull($int("b"), $int(1)));
+
+        assertExpHashCode(
+                ifNull($int("a"), $int(1)),
+                ifNull($int("a"), $int(1)),
+                ifNull($int("a"), $int(2)));
     }
 }