+++ dflib/src/test/java/org/dflib/exp/num/DecimalColumnTest.java
@@ -7,7 +7,7 @@ import org.dflib.DecimalExp;
 import org.dflib.NumExp;
 import org.dflib.Series;
 import org.dflib.StrExp;
-import org.dflib.exp.ExpBaseTest;
+import org.dflib.exp.BaseExpTest;
 import org.dflib.unit.BoolSeriesAsserts;
 import org.dflib.unit.SeriesAsserts;
 import org.junit.jupiter.api.Test;
@@ -19,7 +19,7 @@ import static org.dflib.Exp.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
 
-public class DecimalColumnTest extends ExpBaseTest {
+public class DecimalColumnTest extends BaseExpTest {
 
     @Test
     public void getColumnName() {
@@ -399,13 +399,18 @@ public class DecimalColumnTest extends ExpBaseTest {
     }
 
     @Test
-    public void equalsHashCode() {
-        DecimalExp e1 = $decimal("a");
-        DecimalExp e2 = $decimal("a");
-        DecimalExp e3 = $decimal("a");
-        DecimalExp different = $decimal("b");
+    public void testEquals() {
+        assertExpEquals(
+                $decimal("a"),
+                $decimal("a"),
+                $decimal("b"));
+    }
 
-        assertEqualsContract(e1, e2, e3);
-        assertNotEquals(e1, different);
+    @Test
+    public void testHashCode() {
+        assertExpHashCode(
+                $decimal("a"),
+                $decimal("a"),
+                $decimal("b"));
     }
 }