+++ dflib/src/test/java/org/dflib/exp/num/DoubleBinaryExpTest.java
@@ -1,17 +1,14 @@
 package org.dflib.exp.num;
 
+import org.dflib.exp.BaseExpTest;
 import org.junit.jupiter.api.Test;
 
 import static org.dflib.Exp.$double;
 import static org.dflib.Exp.$int;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.dflib.NumExp;
-import org.dflib.exp.ExpBaseTest;
-
-public class DoubleBinaryExpTest extends ExpBaseTest {
+public class DoubleBinaryExpTest extends BaseExpTest {
 
     @Test
     public void creation() {
@@ -27,13 +24,18 @@ public class DoubleBinaryExpTest extends ExpBaseTest {
     }
 
     @Test
-    public void equalsHashCode() {
-        NumExp<?> e1 = $double(1).add($int(2));
-        NumExp<?> e2 = $double(1).add($int(2));
-        NumExp<?> e3 = $double(1).add($int(2));
-        NumExp<?> different = $double(1).add($int(3));
+    public void testEquals() {
+        assertExpEquals(
+                $double(1).add($int(2)),
+                $double(1).add($int(2)),
+                $double(1).add($int(3)));
+    }
 
-        assertEqualsContract(e1, e2, e3);
-        assertNotEquals(e1, different);
+    @Test
+    public void testHashCode() {
+        assertExpHashCode(
+                $double(1).add($int(2)),
+                $double(1).add($int(2)),
+                $double(1).add($int(3)));
     }
 }