+++ dflib/src/test/java/org/dflib/exp/num/DoubleColumnTest.java
@@ -5,7 +5,7 @@ import org.dflib.DecimalExp;
 import org.dflib.DoubleSeries;
 import org.dflib.NumExp;
 import org.dflib.Series;
-import org.dflib.exp.ExpBaseTest;
+import org.dflib.exp.BaseExpTest;
 import org.dflib.unit.SeriesAsserts;
 import org.junit.jupiter.api.Test;
 
@@ -16,7 +16,7 @@ import static org.dflib.Exp.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
 
-public class DoubleColumnTest extends ExpBaseTest {
+public class DoubleColumnTest extends BaseExpTest {
 
     @Test
     public void getColumnName() {
@@ -188,13 +188,18 @@ public class DoubleColumnTest extends ExpBaseTest {
     }
 
     @Test
-    public void equalsHashCode() {
-        NumExp<?> e1 = $double("a");
-        NumExp<?> e2 = $double("a");
-        NumExp<?> e3 = $double("a");
-        NumExp<?> different = $double("b");
+    public void testEquals() {
+        assertExpEquals(
+                $double("a"),
+                $double("a"),
+                $double("b"));
+    }
 
-        assertEqualsContract(e1, e2, e3);
-        assertNotEquals(e1, different);
+    @Test
+    public void testHashCode() {
+        assertExpHashCode(
+                $double("a"),
+                $double("a"),
+                $double("b"));
     }
 }