+++ dflib/src/test/java/org/dflib/exp/num/FloatColumnTest.java
@@ -6,7 +6,7 @@ import org.dflib.DoubleSeries;
 import org.dflib.FloatSeries;
 import org.dflib.NumExp;
 import org.dflib.Series;
-import org.dflib.exp.ExpBaseTest;
+import org.dflib.exp.BaseExpTest;
 import org.dflib.unit.SeriesAsserts;
 import org.junit.jupiter.api.Test;
 
@@ -16,7 +16,7 @@ import static org.dflib.Exp.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
 
-public class FloatColumnTest extends ExpBaseTest {
+public class FloatColumnTest extends BaseExpTest {
 
     @Test
     public void getColumnName() {
@@ -201,13 +201,18 @@ public class FloatColumnTest extends ExpBaseTest {
     }
 
     @Test
-    public void equalsHashCode() {
-        NumExp<?> e1 = $float("a");
-        NumExp<?> e2 = $float("a");
-        NumExp<?> e3 = $float("a");
-        NumExp<?> different = $float("b");
+    public void testEquals() {
+        assertExpEquals(
+                $float("a"),
+                $float("a"),
+                $float("b"));
+    }
 
-        assertEqualsContract(e1, e2, e3);
-        assertNotEquals(e1, different);
+    @Test
+    public void testHashCode() {
+        assertExpHashCode(
+                $float("a"),
+                $float("a"),
+                $float("b"));
     }
 }