+++ dflib/src/test/java/org/dflib/exp/str/ConcatExpTest.java
@@ -1,28 +1,26 @@
 package org.dflib.exp.str;
 
 import org.dflib.DataFrame;
-import org.dflib.Exp;
 import org.dflib.StrExp;
-import org.dflib.exp.ExpBaseTest;
+import org.dflib.exp.BaseExpTest;
 import org.dflib.unit.SeriesAsserts;
 import org.junit.jupiter.api.Test;
 
 import static org.dflib.Exp.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-public class ConcatExpTest extends ExpBaseTest {
+public class ConcatExpTest extends BaseExpTest {
 
     @Test
     public void toQL() {
-        StrExp exp = Exp.concat($str("a"), $val("X"), $str(0));
+        StrExp exp = concat($str("a"), $val("X"), $str(0));
         assertEquals("concat(a,'X',$str(0))", exp.toQL());
     }
 
     @Test
-    public void concat() {
-        StrExp exp1 = Exp.concat($str("b"), $int("a"));
-        StrExp exp2 = Exp.concat("_", $str("b"), "]");
+    public void test() {
+        StrExp exp1 = concat($str("b"), $int("a"));
+        StrExp exp2 = concat("_", $str("b"), "]");
 
         DataFrame df = DataFrame.foldByRow("a", "b", "c").of(
                 1, "2", "3",
@@ -35,18 +33,49 @@ public class ConcatExpTest extends ExpBaseTest {
     }
 
     @Test
-    public void equalsHashCode() {
-        Exp<?> e1 = Exp.concat($str("a"), $str("b"), $str("c"));
-        Exp<?> e2 = Exp.concat($str("a"), $str("b"), $str("c"));
-        Exp<?> e3 = Exp.concat($str("a"), $str("b"), $str("c"));
+    public void testEquals() {
+        assertExpEquals(
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("x"), $str("b"), $str("c")));
 
-        Exp<?> different1 = Exp.concat($str("x"), $str("b"), $str("c"));
-        Exp<?> different2 = Exp.concat($str("a"), $str("b"));
-        Exp<?> different3 = Exp.concat($str("a"), $str("b"), $int("c"));
+        assertExpEquals(
+                concat($str("a"), $val("b"), $str("c")),
+                concat($str("a"), "b", $str("c")),
+                concat($str("x"), $val("b"), $str("c")));
 
-        assertEqualsContract(e1, e2, e3);
-        assertNotEquals(e1, different1, "Should not equal expression with a different argument");
-        assertNotEquals(e1, different2, "Should not equal expression with fewer arguments");
-        assertNotEquals(e1, different3, "Should not equal expression with different argument types");
+
+        assertExpEquals(
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b")));
+
+        assertExpEquals(
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $int("c")));
+    }
+
+    @Test
+    public void testHashCode() {
+        assertExpHashCode(
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("x"), $str("b"), $str("c")));
+
+        assertExpHashCode(
+                concat($str("a"), $val("b"), $str("c")),
+                concat($str("a"), "b", $str("c")),
+                concat($str("x"), $val("b"), $str("c")));
+
+        assertExpHashCode(
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b")));
+
+        assertExpHashCode(
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $str("c")),
+                concat($str("a"), $str("b"), $int("c")));
     }
 }